style(dropdown-web): rename type names to Combobox
