+++ packages/pluggableWidgets/dropdown-web/src/Combobox.tsx
@@ -1,9 +1,9 @@
 import { createElement, ReactElement } from "react";
 
-import { DropdownContainerProps } from "../typings/DropdownProps";
+import { ComboboxContainerProps } from "../typings/ComboboxProps";
 import { Dropdown as DropdownComponent } from "./components/Dropdown";
 import "./ui/Dropdown.scss";
 
-export default function Dropdown(props: DropdownContainerProps): ReactElement {
+export default function Dropdown(props: ComboboxContainerProps): ReactElement {
     return <DropdownComponent {...props} />;
 }
