+++ packages/pluggableWidgets/dropdown-web/src/components/Dropdown.tsx
@@ -2,14 +2,14 @@ import classNames from "classnames";
 import Downshift from "downshift";
 import { createElement, ReactElement, useRef } from "react";
 import { useActionEvents } from "../hooks/useActionEvents";
-import { DropdownContainerProps } from "../../typings/DropdownProps";
+import { ComboboxContainerProps } from "../../typings/ComboboxProps";
 import { ClearButton, DownArrow } from "../assets/icons";
 import { useDownshiftProps } from "../hooks/useDownshiftProps";
 import { useGetSelector } from "../hooks/useGetSelector";
 import { DropdownMenu } from "./DropdownMenu";
 import { Placeholder } from "./Placeholder";
 
-export function Dropdown(props: DropdownContainerProps): ReactElement {
+export function Dropdown(props: ComboboxContainerProps): ReactElement {
     const inputRef = useRef<HTMLInputElement>(null);
     const dropdownRef = useRef<HTMLInputElement>(null);
     const selector = useGetSelector(props);
