+++ packages/pluggableWidgets/dropdown-web/src/helpers/getSelector.ts
@@ -1,9 +1,9 @@
-import { DropdownContainerProps } from "../../typings/DropdownProps";
+import { ComboboxContainerProps } from "../../typings/ComboboxProps";
 import { AssociationSingleSelector } from "./Association/AssociationSingleSelector";
 import { EnumBooleanSingleSelector } from "./EnumBool/EnumBoolSingleSelector";
 import { SingleSelector } from "./types";
 
-export function getSelector(props: DropdownContainerProps): SingleSelector {
+export function getSelector(props: ComboboxContainerProps): SingleSelector {
     if (props.optionsSourceType === "enumeration" || props.optionsSourceType === "boolean") {
         return new EnumBooleanSingleSelector();
     } else if (props.optionsSourceType === "association") {
