+++ packages/pluggableWidgets/dropdown-web/src/hooks/useGetSelector.ts
@@ -1,9 +1,9 @@
 import { useRef } from "react";
 import { getSelector } from "../helpers/getSelector";
 import { SingleSelector } from "../helpers/types";
-import { DropdownContainerProps } from "../../typings/DropdownProps";
+import { ComboboxContainerProps } from "../../typings/ComboboxProps";
 
-export function useGetSelector(props: DropdownContainerProps): SingleSelector {
+export function useGetSelector(props: ComboboxContainerProps): SingleSelector {
     const selectorRef = useRef<SingleSelector | undefined>(undefined);
     if (!selectorRef.current) {
         selectorRef.current = getSelector(props);
